requirements.dev.txt
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distlib==0.4.0
django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
docker==7.1.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
filelock==3.19.1
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
identify==2.6.14
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy-extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pre-commit==4.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
pyjwt==2.10.1
python-dotenv==1.1.1
pytz==2025.2
# pywin32==311 윈도우 전용 (리눅스, 맥에서는 설치 안함)
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
ruff==0.13.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
